+++ NEWS.md
@@ -1,3 +1,9 @@
+# jstable 0.7.3
+
+## Update
+
+* Add **R-squared** to `glmshow.display`
+
 # jstable 0.7.2
 
 ## Bug fixes